chore: move configuration modification to the Loader
